requirements.txt
anthropic>=0.42.0
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.5.0
rich>=13.7.0
pytest>=7.4.0
